+++ src/core/context/index.ts
@@ -1,2 +1,0 @@
-export * from "./ContextNode";
-export * from "./ContextTree";