+++ src/core/diff/index.ts
@@ -1,3 +1,0 @@
-export * from "./DiffPoint";
-export * from "./DiffViewer";
-export * from "./DiffSummary";