+++ src/core/plain/index.ts
@@ -1,2 +1,0 @@
-export * from "./HTMLAdapter";
-export * from "./HTMLNode";