+++ src/implementation/diff/viewer/index.ts
@@ -1,6 +1,0 @@
-export * from "./AbstractDiffViewer";
-export * from "./ComparingBasedDiffViewer";
-export * from "./TreeHierarchyDiffViewer";
-export * from "./NodeMutationDiffViewer";
-export * from "./SubtreeShapeDiffViewer";
-export * from "./CompositeDiffViewer";